Added some web server features, added model query for plugins ===============================================================
System Information:
Engine Version: Molecule Ver 2.7.2.20250820
Library Version: Ver 3.4.2.20250804
ABI: Ver 1.5.0.20250804
Common: Ver 3.3.0.20250804

Hosts:
Interactive Console Ver 1.0.8.20250812
Pipe Child Ver 1.2.0.20250812
Headless Host Ver 1.1.1.20250812

Storages:
File System Storage Ver 1.4.1.20250812

Firmware:
Windows Ver 2.3.0.20250804
Linux Ver 2.1.1.20250804

Tested Compilers:
MSVC 14.44.35207
GCC 11.4.0

Tested OSes:
Linux Ver:6.8.0-60-generic [x86_64]
Windows Workstation 10.0.26100.4946 [x64]

+++ protocols/http/rx_web_socket_mapping.cpp
@@ -441,7 +441,7 @@ bool rx_web_socket_endpoint::is_header_done ()
     {
         return false; // Header is not complete
 	}
-	int mask_size = (header_buffer_[1] & 0x80) ? 4 : 0; // Check if mask is present
+	size_t mask_size = (header_buffer_[1] & 0x80) ? 4 : 0; // Check if mask is present
     switch(header_buffer_[1]& 0x7F) // Mask the first bit to get payload length
     {
         case 126: // Extended payload length (16 bits)
